Updated Linear Regression Cases
Updated the way in which linear regression would be computed so that a fi for a logarithmic trend would always* come after a fit for an exponential trend. *: Added overflow handling which should prevent some (but not all) overflow errors when computing exp(n) for a population n at time step t.